+++ plutus-extra/CHANGELOG.md
@@ -4,9 +4,12 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## Unreleased
 
+## 4.1 -- 2021-12-14
+
 ### Added
 
-- PlutusTx.Data.Extra: `toDatumHash` to get `DatumHash`
+- PlutusTx.Data.Extra: `toDatumHash` to get `DatumHash` of a value
+- PlutusTx.Set: `Data.OpenApi.Internal.Schema.ToSchema` instance for `Set`
 
 ### Changed